refactor(buffer): extract save/write-recipe cluster to save.rs
Moves SudoSaveRequired, WriteRecipe + RecipeAction, and 11 save-
helper methods (should_use_inplace_write, build_write_recipe,
create_temp_file, create_recovery_temp_file, inplace_recovery_meta_path,
write_inplace_recovery_meta, save_with_inplace_write,
write_data_inplace, stream_file_to_writer, write_recipe_to_file,
make_sudo_error) out of impl TextBuffer in mod.rs and into
model/buffer/save.rs as free functions taking &Arc<dyn FileSystem>,
&PieceTree, &[StringBuffer], &BufferFormat, &BufferFileKind,
&Persistence as needed.

Orchestrators (save, save_to_file, finalize_save,
finalize_external_save, consolidate_*) remain on impl TextBuffer
in mod.rs. save_to_file now destructures fs via
self.persistence.fs() and calls save::build_write_recipe,
save::save_with_inplace_write, etc.

A thin delegator build_write_recipe(&self) remains on TextBuffer
for the property test suite. WriteRecipe and RecipeAction are
re-exported as pub(crate) for test access.

mod.rs drops from 4,384 → 3,892 (−492 lines). save.rs is 565 lines.
127 buffer tests green.
