+++ crates/fresh-editor/src/model/buffer/mod.rs
@@ -2,7 +2,7 @@
 /// Architecture where the tree is the single source of truth for text and line information
 use crate::model::encoding;
 use crate::model::filesystem::{
-    FileMetadata, FileSearchCursor, FileSearchOptions, FileSystem, WriteOp,
+    FileSearchCursor, FileSearchOptions, FileSystem,
 };
 use crate::model::piece_tree::{
     BufferData, BufferLocation, Cursor, PieceInfo, PieceRangeIter, PieceTree, PieceView, Position,
@@ -12,7 +12,8 @@ use crate::model::piece_tree_diff::PieceTreeDiff;
 use crate::primitives::grapheme;
 use anyhow::{Context, Result};
 use regex::bytes::Regex;
-use std::io::{self, Write};
+use std::io;
+
 use std::ops::Range;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
@@ -23,39 +24,14 @@ pub use encoding::Encoding;
 pub mod file_kind;
 pub mod format;
 pub mod persistence;
+pub mod save;
 pub use file_kind::BufferFileKind;
 pub use format::{BufferFormat, LineEnding};
 pub use persistence::Persistence;
+pub use save::SudoSaveRequired;
+#[cfg(test)]
+pub(crate) use save::{RecipeAction, WriteRecipe};
 
-/// Error returned when a file save operation requires elevated privileges.
-///
-/// This error contains all the information needed to perform the save via sudo
-/// in a single operation, preserving original file ownership and permissions.
-#[derive(Debug, Clone, PartialEq)]
-pub struct SudoSaveRequired {
-    /// Path to the temporary file containing the new content
-    pub temp_path: PathBuf,
-    /// Destination path where the file should be saved
-    pub dest_path: PathBuf,
-    /// Original file owner (UID)
-    pub uid: u32,
-    /// Original file group (GID)
-    pub gid: u32,
-    /// Original file permissions (mode)
-    pub mode: u32,
-}
-
-impl std::fmt::Display for SudoSaveRequired {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "Permission denied saving to {}. Use sudo to complete the operation.",
-            self.dest_path.display()
-        )
-    }
-}
-
-impl std::error::Error for SudoSaveRequired {}
 
 /// Error returned when a large file has a non-resynchronizable encoding
 /// and requires user confirmation before loading the entire file into memory.
@@ -191,61 +167,7 @@ impl Default for BufferConfig {
 
 /// Line ending format used in the file
 
-/// A write recipe built from the piece tree for saving
-struct WriteRecipe {
-    /// The source file path for Copy operations (if any)
-    src_path: Option<PathBuf>,
-    /// Data chunks for Insert operations (owned to avoid lifetime issues)
-    insert_data: Vec<Vec<u8>>,
-    /// Sequence of actions to build the output file
-    actions: Vec<RecipeAction>,
-}
 
-/// An action in a write recipe
-#[derive(Debug, Clone, Copy)]
-enum RecipeAction {
-    /// Copy bytes from source file at offset
-    Copy { offset: u64, len: u64 },
-    /// Insert data from insert_data[index]
-    Insert { index: usize },
-}
-
-impl WriteRecipe {
-    /// Convert the recipe to WriteOp slice for use with filesystem write_patched
-    fn to_write_ops(&self) -> Vec<WriteOp<'_>> {
-        self.actions
-            .iter()
-            .map(|action| match action {
-                RecipeAction::Copy { offset, len } => WriteOp::Copy {
-                    offset: *offset,
-                    len: *len,
-                },
-                RecipeAction::Insert { index } => WriteOp::Insert {
-                    data: &self.insert_data[*index],
-                },
-            })
-            .collect()
-    }
-
-    /// Check if this recipe has any Copy operations
-    fn has_copy_ops(&self) -> bool {
-        self.actions
-            .iter()
-            .any(|a| matches!(a, RecipeAction::Copy { .. }))
-    }
-
-    /// Flatten all Insert operations into a single buffer.
-    /// Only valid when has_copy_ops() returns false.
-    fn flatten_inserts(&self) -> Vec<u8> {
-        let mut result = Vec::new();
-        for action in &self.actions {
-            if let RecipeAction::Insert { index } = action {
-                result.extend_from_slice(&self.insert_data[*index]);
-            }
-        }
-        result
-    }
-}
 
 /// Represents a line number (simplified for new implementation)
 /// Legacy enum kept for backwards compatibility - always Absolute now
@@ -778,262 +700,22 @@ impl TextBuffer {
         }
     }
 
-    /// Check if we should use in-place writing to preserve file ownership.
-    /// Returns true if the file exists and is owned by a different user.
-    /// On Unix, only root or the file owner can change file ownership with chown.
-    /// When the current user is not the file owner, using atomic write (temp file + rename)
-    /// would change the file's ownership to the current user. To preserve ownership,
-    /// we must write directly to the existing file instead.
-    fn should_use_inplace_write(&self, dest_path: &Path) -> bool {
-        !self.persistence.fs().is_owner(dest_path)
-    }
-
     /// Build a write recipe from the piece tree for saving.
     ///
-    /// This creates a recipe of Copy and Insert operations that can reconstruct
-    /// the buffer content. Copy operations reference unchanged regions in the
-    /// source file, while Insert operations contain new/modified data.
-    ///
-    /// # Returns
-    /// A WriteRecipe with the source path, insert data, and sequence of actions.
-    fn build_write_recipe(&self) -> io::Result<WriteRecipe> {
-        let total = self.total_bytes();
-
-        // Determine the source file for Copy operations (if any)
-        // We can only use Copy if:
-        // 1. We have a source file path
-        // 2. The source file exists
-        // 3. No line ending conversion is needed
-        // 4. No encoding conversion is needed
-        let needs_line_ending_conversion = self.format.line_ending_changed_since_load();
-        // We need encoding conversion if:
-        // - NOT a binary file (binary files preserve raw bytes), AND
-        // - Either the encoding changed from the original, OR
-        // - The target encoding isn't plain UTF-8/ASCII (since internal storage is UTF-8)
-        // For example: UTF-8 BOM files are stored as UTF-8, so we need to add BOM on save
-        let needs_encoding_conversion = !self.file_kind.is_binary()
-            && (self.format.encoding_changed_since_load()
-                || !matches!(
-                    self.format.encoding(),
-                    Encoding::Utf8 | Encoding::Ascii
-                ));
-        let needs_conversion = needs_line_ending_conversion || needs_encoding_conversion;
-
-        let src_path_for_copy: Option<&Path> = if needs_conversion {
-            None
-        } else {
-            self.persistence.file_path().filter(|p| self.persistence.fs().exists(p))
-        };
-        let target_ending = self.format.line_ending();
-        let target_encoding = self.format.encoding();
-
-        let mut insert_data: Vec<Vec<u8>> = Vec::new();
-        let mut actions: Vec<RecipeAction> = Vec::new();
-
-        // Add BOM as the first piece if the target encoding has one
-        if let Some(bom) = target_encoding.bom_bytes() {
-            insert_data.push(bom.to_vec());
-            actions.push(RecipeAction::Insert { index: 0 });
-        }
-
-        for piece_view in self.piece_tree.iter_pieces_in_range(0, total) {
-            let buffer_id = piece_view.location.buffer_id();
-            let buffer = self.buffers.get(buffer_id).ok_or_else(|| {
-                io::Error::new(
-                    io::ErrorKind::InvalidData,
-                    format!("Buffer {} not found", buffer_id),
-                )
-            })?;
-
-            match &buffer.data {
-                // Unloaded buffer: can use Copy if same source file, else load and send
-                BufferData::Unloaded {
-                    file_path,
-                    file_offset,
-                    ..
-                } => {
-                    // Can only use Copy if:
-                    // - This is a Stored piece (original file content)
-                    // - We have a valid source for copying
-                    // - This buffer is from that source
-                    // - No line ending or encoding conversion needed
-                    let can_copy = matches!(piece_view.location, BufferLocation::Stored(_))
-                        && src_path_for_copy.is_some_and(|src| file_path == src);
-
-                    if can_copy {
-                        let src_offset = (*file_offset + piece_view.buffer_offset) as u64;
-                        actions.push(RecipeAction::Copy {
-                            offset: src_offset,
-                            len: piece_view.bytes as u64,
-                        });
-                        continue;
-                    }
-
-                    // Need to load and send this unloaded region
-                    // This happens when: different source file, or conversion needed
-                    let data = self.persistence.fs().read_range(
-                        file_path,
-                        (*file_offset + piece_view.buffer_offset) as u64,
-                        piece_view.bytes,
-                    )?;
-
-                    let data = if needs_line_ending_conversion {
-                        format::convert_line_endings_to(&data, target_ending)
-                    } else {
-                        data
-                    };
-
-                    // Convert encoding if needed
-                    let data = if needs_encoding_conversion {
-                        format::convert_to_encoding(&data, target_encoding)
-                    } else {
-                        data
-                    };
-
-                    let index = insert_data.len();
-                    insert_data.push(data);
-                    actions.push(RecipeAction::Insert { index });
-                }
-
-                // Loaded data: send as Insert
-                BufferData::Loaded { data, .. } => {
-                    let start = piece_view.buffer_offset;
-                    let end = start + piece_view.bytes;
-                    let chunk = &data[start..end];
-
-                    let chunk = if needs_line_ending_conversion {
-                        format::convert_line_endings_to(chunk, target_ending)
-                    } else {
-                        chunk.to_vec()
-                    };
-
-                    // Convert encoding if needed
-                    let chunk = if needs_encoding_conversion {
-                        format::convert_to_encoding(&chunk, target_encoding)
-                    } else {
-                        chunk
-                    };
-
-                    let index = insert_data.len();
-                    insert_data.push(chunk);
-                    actions.push(RecipeAction::Insert { index });
-                }
-            }
-        }
-
-        Ok(WriteRecipe {
-            src_path: src_path_for_copy.map(|p| p.to_path_buf()),
-            insert_data,
-            actions,
-        })
+    /// Delegates to `save::build_write_recipe`.
+    #[cfg(test)]
+    pub(crate) fn build_write_recipe(&self) -> io::Result<WriteRecipe> {
+        save::build_write_recipe(
+            &self.piece_tree,
+            &self.buffers,
+            &self.format,
+            &self.file_kind,
+            &self.persistence,
+        )
     }
 
-    /// Create a temporary file for saving.
-    ///
-    /// Tries to create the file in the same directory as the destination file first
-    /// to allow for an atomic rename. If that fails (e.g., due to directory permissions),
-    /// falls back to the system temporary directory.
-    fn create_temp_file(
-        &self,
-        dest_path: &Path,
-    ) -> io::Result<(PathBuf, Box<dyn crate::model::filesystem::FileWriter>)> {
-        // Try creating in same directory first
-        let same_dir_temp = self.persistence.fs().temp_path_for(dest_path);
-        match self.persistence.fs().create_file(&same_dir_temp) {
-            Ok(file) => Ok((same_dir_temp, file)),
-            Err(e) if e.kind() == io::ErrorKind::PermissionDenied => {
-                // Fallback to system temp directory
-                let temp_path = self.persistence.fs().unique_temp_path(dest_path);
-                let file = self.persistence.fs().create_file(&temp_path)?;
-                Ok((temp_path, file))
-            }
-            Err(e) => Err(e),
-        }
-    }
 
-    /// Create a temporary file in the recovery directory for in-place writes.
-    /// This allows recovery if a crash occurs during the in-place write operation.
-    fn create_recovery_temp_file(
-        &self,
-        dest_path: &Path,
-    ) -> io::Result<(PathBuf, Box<dyn crate::model::filesystem::FileWriter>)> {
-        // Get recovery directory: $XDG_DATA_HOME/fresh/recovery or ~/.local/share/fresh/recovery
-        let recovery_dir = crate::input::input_history::get_data_dir()
-            .map(|d| d.join("recovery"))
-            .unwrap_or_else(|_| std::env::temp_dir());
 
-        // Ensure directory exists
-        self.persistence.fs().create_dir_all(&recovery_dir)?;
-
-        // Create unique filename based on destination file and timestamp
-        let file_name = dest_path
-            .file_name()
-            .unwrap_or_else(|| std::ffi::OsStr::new("fresh-save"));
-        let timestamp = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.as_nanos())
-            .unwrap_or(0);
-        let pid = std::process::id();
-
-        let temp_name = format!(
-            ".inplace-{}-{}-{}.tmp",
-            file_name.to_string_lossy(),
-            pid,
-            timestamp
-        );
-        let temp_path = recovery_dir.join(temp_name);
-
-        let file = self.persistence.fs().create_file(&temp_path)?;
-        Ok((temp_path, file))
-    }
-
-    /// Get the path for in-place write recovery metadata.
-    /// Uses the same recovery directory as temp files.
-    fn inplace_recovery_meta_path(&self, dest_path: &Path) -> PathBuf {
-        let recovery_dir = crate::input::input_history::get_data_dir()
-            .map(|d| d.join("recovery"))
-            .unwrap_or_else(|_| std::env::temp_dir());
-
-        let hash = crate::services::recovery::path_hash(dest_path);
-        recovery_dir.join(format!("{}.inplace.json", hash))
-    }
-
-    /// Write in-place recovery metadata using self.persistence.fs().
-    /// This is called before the dangerous streaming step so we can recover on crash.
-    fn write_inplace_recovery_meta(
-        &self,
-        meta_path: &Path,
-        dest_path: &Path,
-        temp_path: &Path,
-        original_metadata: &Option<FileMetadata>,
-    ) -> io::Result<()> {
-        #[cfg(unix)]
-        let (uid, gid, mode) = original_metadata
-            .as_ref()
-            .map(|m| {
-                (
-                    m.uid.unwrap_or(0),
-                    m.gid.unwrap_or(0),
-                    m.permissions.as_ref().map(|p| p.mode()).unwrap_or(0o644),
-                )
-            })
-            .unwrap_or((0, 0, 0o644));
-        #[cfg(not(unix))]
-        let (uid, gid, mode) = (0u32, 0u32, 0o644u32);
-
-        let recovery = crate::services::recovery::InplaceWriteRecovery::new(
-            dest_path.to_path_buf(),
-            temp_path.to_path_buf(),
-            uid,
-            gid,
-            mode,
-        );
-
-        let json = serde_json::to_string_pretty(&recovery)
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-
-        self.persistence.fs().write_file(meta_path, json.as_bytes())
-    }
 
     /// Save the buffer to a specific file
     ///
@@ -1061,204 +743,63 @@ impl TextBuffer {
         }
 
         // Build the write recipe (unified for all filesystem types)
-        let recipe = self.build_write_recipe()?;
+        let recipe = save::build_write_recipe(
+            &self.piece_tree,
+            &self.buffers,
+            &self.format,
+            &self.file_kind,
+            &self.persistence,
+        )?;
         let ops = recipe.to_write_ops();
 
         // Check if we need in-place writing to preserve file ownership (local only)
         // Remote filesystems handle this differently
-        let is_local = self.persistence.fs().remote_connection_info().is_none();
-        let use_inplace = is_local && self.should_use_inplace_write(dest_path);
+        let fs = self.persistence.fs();
+        let is_local = fs.remote_connection_info().is_none();
+        let use_inplace = is_local && save::should_use_inplace_write(fs, dest_path);
 
         if use_inplace {
             // In-place write: write directly to preserve ownership
-            self.save_with_inplace_write(dest_path, &recipe)?;
+            save::save_with_inplace_write(fs, dest_path, &recipe)?;
         } else if !recipe.has_copy_ops() && !is_local {
             // Remote with no Copy ops: use write_file directly (more efficient)
             let data = recipe.flatten_inserts();
-            self.persistence.fs().write_file(dest_path, &data)?;
+            fs.write_file(dest_path, &data)?;
         } else if is_local {
             // Local: use write_file or write_patched with sudo fallback
             let write_result = if !recipe.has_copy_ops() {
                 let data = recipe.flatten_inserts();
-                self.persistence.fs().write_file(dest_path, &data)
+                fs.write_file(dest_path, &data)
             } else {
                 let src_for_patch = recipe.src_path.as_deref().unwrap_or(dest_path);
-                self.persistence.fs().write_patched(src_for_patch, dest_path, &ops)
+                fs.write_patched(src_for_patch, dest_path, &ops)
             };
 
             if let Err(e) = write_result {
                 if e.kind() == io::ErrorKind::PermissionDenied {
                     // Create temp file and return sudo error
-                    let original_metadata = self.persistence.fs().metadata_if_exists(dest_path);
-                    let (temp_path, mut temp_file) = self.create_temp_file(dest_path)?;
-                    self.write_recipe_to_file(&mut temp_file, &recipe)?;
+                    let original_metadata = fs.metadata_if_exists(dest_path);
+                    let (temp_path, mut temp_file) = save::create_temp_file(fs, dest_path)?;
+                    save::write_recipe_to_file(fs, &mut temp_file, &recipe)?;
                     temp_file.sync_all()?;
                     drop(temp_file);
-                    return Err(self.make_sudo_error(temp_path, dest_path, original_metadata));
+                    return Err(save::make_sudo_error(temp_path, dest_path, original_metadata));
                 }
                 return Err(e.into());
             }
         } else {
             // Remote with Copy ops: use write_patched
             let src_for_patch = recipe.src_path.as_deref().unwrap_or(dest_path);
-            self.persistence.fs().write_patched(src_for_patch, dest_path, &ops)?;
+            fs.write_patched(src_for_patch, dest_path, &ops)?;
         }
 
         self.finalize_save(dest_path)?;
         Ok(())
     }
 
-    /// Write using in-place mode to preserve file ownership.
-    ///
-    /// This is used when the file is owned by a different user and we need
-    /// to write directly to the existing file to preserve its ownership.
-    ///
-    /// The approach:
-    /// 1. Write the recipe to a temp file first (reads from original, writes to temp)
-    /// 2. Stream the temp file content to the destination file (truncates and writes)
-    /// 3. Delete the temp file
-    ///
-    /// This avoids the bug where truncating the destination before reading Copy chunks
-    /// would corrupt the file. It also works for huge files since we stream in chunks.
-    fn save_with_inplace_write(
-        &self,
-        dest_path: &Path,
-        recipe: &WriteRecipe,
-    ) -> anyhow::Result<()> {
-        let original_metadata = self.persistence.fs().metadata_if_exists(dest_path);
 
-        // Optimization: if no Copy ops, we can write directly without a temp file
-        // (same as the non-inplace path for small files)
-        if !recipe.has_copy_ops() {
-            let data = recipe.flatten_inserts();
-            return self.write_data_inplace(dest_path, &data, original_metadata);
-        }
 
-        // Step 1: Write recipe to a temp file in the recovery directory
-        // This reads Copy chunks from the original file (still intact) and writes to temp.
-        // Using the recovery directory allows crash recovery if the operation fails.
-        let (temp_path, mut temp_file) = self.create_recovery_temp_file(dest_path)?;
-        if let Err(e) = self.write_recipe_to_file(&mut temp_file, recipe) {
-            // Best-effort cleanup of temp file on write failure
-            #[allow(clippy::let_underscore_must_use)]
-            let _ = self.persistence.fs().remove_file(&temp_path);
-            return Err(e.into());
-        }
-        temp_file.sync_all()?;
-        drop(temp_file);
 
-        // Step 1.5: Save recovery metadata before the dangerous step
-        // If we crash during step 2, this metadata + temp file allows recovery
-        let recovery_meta_path = self.inplace_recovery_meta_path(dest_path);
-        // Best effort - don't fail the save if we can't write recovery metadata
-        #[allow(clippy::let_underscore_must_use)]
-        let _ = self.write_inplace_recovery_meta(
-            &recovery_meta_path,
-            dest_path,
-            &temp_path,
-            &original_metadata,
-        );
-
-        // Step 2: Stream temp file content to destination
-        // Now it's safe to truncate the destination since all data is in temp
-        match self.persistence.fs().open_file_for_write(dest_path) {
-            Ok(mut out_file) => {
-                if let Err(e) = self.stream_file_to_writer(&temp_path, &mut out_file) {
-                    // Don't delete temp file or recovery metadata - allow recovery
-                    return Err(e.into());
-                }
-                out_file.sync_all()?;
-                // Success! Clean up temp file and recovery metadata (best-effort)
-                #[allow(clippy::let_underscore_must_use)]
-                let _ = self.persistence.fs().remove_file(&temp_path);
-                #[allow(clippy::let_underscore_must_use)]
-                let _ = self.persistence.fs().remove_file(&recovery_meta_path);
-                Ok(())
-            }
-            Err(e) if e.kind() == io::ErrorKind::PermissionDenied => {
-                // Can't write to destination - trigger sudo fallback
-                // Keep temp file for sudo to use, clean up recovery metadata (best-effort)
-                #[allow(clippy::let_underscore_must_use)]
-                let _ = self.persistence.fs().remove_file(&recovery_meta_path);
-                Err(self.make_sudo_error(temp_path, dest_path, original_metadata))
-            }
-            Err(e) => {
-                // Don't delete temp file or recovery metadata - allow recovery
-                Err(e.into())
-            }
-        }
-    }
-
-    /// Write data directly to a file in-place, with sudo fallback on permission denied.
-    fn write_data_inplace(
-        &self,
-        dest_path: &Path,
-        data: &[u8],
-        original_metadata: Option<FileMetadata>,
-    ) -> anyhow::Result<()> {
-        match self.persistence.fs().open_file_for_write(dest_path) {
-            Ok(mut out_file) => {
-                out_file.write_all(data)?;
-                out_file.sync_all()?;
-                Ok(())
-            }
-            Err(e) if e.kind() == io::ErrorKind::PermissionDenied => {
-                // Create temp file for sudo fallback
-                let (temp_path, mut temp_file) = self.create_temp_file(dest_path)?;
-                temp_file.write_all(data)?;
-                temp_file.sync_all()?;
-                drop(temp_file);
-                Err(self.make_sudo_error(temp_path, dest_path, original_metadata))
-            }
-            Err(e) => Err(e.into()),
-        }
-    }
-
-    /// Stream a file's content to a writer in chunks to avoid memory issues with large files.
-    fn stream_file_to_writer(
-        &self,
-        src_path: &Path,
-        out_file: &mut Box<dyn crate::model::filesystem::FileWriter>,
-    ) -> io::Result<()> {
-        const CHUNK_SIZE: usize = 1024 * 1024; // 1MB chunks
-
-        let file_size = self.persistence.fs().metadata(src_path)?.size;
-        let mut offset = 0u64;
-
-        while offset < file_size {
-            let remaining = file_size - offset;
-            let chunk_len = std::cmp::min(remaining, CHUNK_SIZE as u64) as usize;
-            let chunk = self.persistence.fs().read_range(src_path, offset, chunk_len)?;
-            out_file.write_all(&chunk)?;
-            offset += chunk_len as u64;
-        }
-
-        Ok(())
-    }
-
-    /// Write the recipe content to a file writer.
-    fn write_recipe_to_file(
-        &self,
-        out_file: &mut Box<dyn crate::model::filesystem::FileWriter>,
-        recipe: &WriteRecipe,
-    ) -> io::Result<()> {
-        for action in &recipe.actions {
-            match action {
-                RecipeAction::Copy { offset, len } => {
-                    // Read from source and write to output
-                    let src_path = recipe.src_path.as_ref().ok_or_else(|| {
-                        io::Error::new(io::ErrorKind::InvalidData, "Copy action without source")
-                    })?;
-                    let data = self.persistence.fs().read_range(src_path, *offset, *len as usize)?;
-                    out_file.write_all(&data)?;
-                }
-                RecipeAction::Insert { index } => {
-                    out_file.write_all(&recipe.insert_data[*index])?;
-                }
-            }
-        }
-        Ok(())
-    }
 
     /// Finalize save state after successful write.
     fn finalize_save(&mut self, dest_path: &Path) -> anyhow::Result<()> {
@@ -1368,39 +909,6 @@ impl TextBuffer {
         }
     }
 
-    /// Internal helper to create a SudoSaveRequired error.
-    fn make_sudo_error(
-        &self,
-        temp_path: PathBuf,
-        dest_path: &Path,
-        original_metadata: Option<FileMetadata>,
-    ) -> anyhow::Error {
-        #[cfg(unix)]
-        let (uid, gid, mode) = if let Some(ref meta) = original_metadata {
-            (
-                meta.uid.unwrap_or(0),
-                meta.gid.unwrap_or(0),
-                meta.permissions
-                    .as_ref()
-                    .map(|p| p.mode() & 0o7777)
-                    .unwrap_or(0),
-            )
-        } else {
-            (0, 0, 0)
-        };
-        #[cfg(not(unix))]
-        let (uid, gid, mode) = (0u32, 0u32, 0u32);
-
-        let _ = original_metadata; // suppress unused warning on non-Unix
-
-        anyhow::anyhow!(SudoSaveRequired {
-            temp_path,
-            dest_path: dest_path.to_path_buf(),
-            uid,
-            gid,
-            mode,
-        })
-    }
 
     /// Get the total number of bytes in the document
     pub fn total_bytes(&self) -> usize {
